requirements.txt
astral==3.2
Babel==2.11.0
cachetools==5.2.0
cairocffi==1.4.0
CairoSVG==2.5.2
caldav==0.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
cssselect2==0.7.0
DateTime==4.3
defusedxml==0.7.1
drawsvg==2.3.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
httplib2==0.20.4
humanize==4.6.0
icalendar==4.0.9
icalevents==0.1.27
idna==3.4
inky==1.5.0
lxml==4.9.1
msal==1.20.0
oauthlib==3.2.2
Pillow==9.3.0
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
RPi.GPIO==0.7.1
rsa==4.9
six==1.16.0
spidev==3.6
tinycss2==1.2.1
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
vobject==0.9.6.1
webencodings==0.5.1
zope.interface==5.5.1
